Refactor ssr server structure to be more future-proof
