+++ packages/react-union-ssr-server/src/middleware/responseCaptureMiddleware.js
@@ -1,29 +1,6 @@
 const { noop } = require('ramda-extension');
 
 /**
- * `createHandler` property for WebpackHotServerMiddleware
- */
-const createHotServerHandler = (error, handleRequest) => (req, res, next) => {
-	if (error) {
-		return next(error);
-	}
-
-	if (req.url !== '/index.html' && req.url !== '/') {
-		return next();
-	}
-
-	try {
-		next();
-		// NOTE: We are essentially emulating the `bodyParser.text()` middleware here.
-		// `res.body` is populated using the `responseCaptureMiddleware` in the dev server.
-		req.body = res.body;
-		handleRequest(req, res, next);
-	} catch (error) {
-		next(error);
-	}
-};
-
-/**
  * Because browserSync only uses Connect and not Express or any other framework,
  * we need to gather the response body ourselves. Because `res.end` is called under the hood
  * to send the response to the client, we intercept the call and instead use `forceEnd`
@@ -57,7 +34,4 @@ const responseCaptureMiddleware = (req, res, next) => {
 	next();
 };
 
-module.exports = {
-	createHotServerHandler,
-	responseCaptureMiddleware,
-};
+module.exports = responseCaptureMiddleware;
