+++ packages/react-union-ssr-server/src/routes/health.js
@@ -1,0 +1,5 @@
+module.exports = () => (req, res) => {
+	res.writeHead(200);
+	// TODO: Perhaps provide more info, such as the uptime, last call, etc.
+	res.end('Running!');
+};
