+++ packages/react-union-ssr-server/src/server.js
@@ -1,52 +1,31 @@
 const connect = require('connect');
 const http = require('http');
 const bodyParser = require('body-parser');
+const R = require('ramda');
 
-const makeContentRenderer = require('./core');
-const { getPort, printUsage } = require('./utils');
+const { getPort, printUsage } = require('./utils/cli');
+const makeRender = require('./render');
+const makeRenderHandler = require('./routes/render');
+const makeHealthHandler = require('./routes/health');
 
 // TODO: maybe add some options as the second parameter
 const startServer = applicationHandler => {
-	const app = connect();
-	const renderContent = makeContentRenderer(applicationHandler);
-
-	app.use(bodyParser.text());
-
-	// NOTE: `options` will be provided by WebpackHotServerMiddleware iff we are running a dev server
-	// The signature matches the `createHandler` property
-	const makeHandleRequest = options => async (req, res, next) => {
-		try {
-			const content = await renderContent(req.body, options, { req, res });
-
-			// NOTE: `res.useForce` is true iff we are running a dev server
-			// TODO: more HTTP headers
-			if (res.useForce) {
-				res.statusMessage = 'OK';
-				res.forceWriteHead(200);
-				res.forceEnd(content);
-			} else {
-				res.writeHead(200, { 'Content-Type': 'application/html' });
-				res.end(content);
-			}
-		} catch (error) {
-			// TODO: Modify HTTP response based on types of errors (bad portal data, React error, etc.)
-			next(error);
-		}
-	};
-
+	const render = makeRender(applicationHandler);
 	// NOTE: this global is undefined iff we are running a dev server, otherwise, it is provided
 	// in the `react-union-scripts` build script
 	if (!global.CLIENT_STATS) {
-		return makeHandleRequest;
+		return R.partial(makeRenderHandler, [render]);
 	}
 
-	app.use('/health', (req, res) => {
-		res.writeHead(200);
-		// TODO: Perhaps provide more info, such as the uptime, last call, etc.
-		res.end('Running!');
-	});
+	const options = {
+		clientStats: global.CLIENT_STATS,
+		isPrebuilt: true,
+	};
 
-	app.use(makeHandleRequest({ clientStats: global.CLIENT_STATS, isPrebuilt: true }));
+	const app = connect();
+	app.use(bodyParser.text());
+	app.use('/', makeRenderHandler(render, options));
+	app.use('/health', makeHealthHandler());
 
 	const server = http.createServer(app);
 	const port = getPort();
