+++ packages/react-union-ssr-server/src/utils/cli.js
@@ -1,31 +1,4 @@
 const R = require('ramda');
-const { rejectNil } = require('ramda-extension');
-
-const hoistComponentStatics = R.compose(
-	R.forEach(R.call(R.__)),
-	rejectNil,
-	R.map(R.path(['component', 'preloadWeak']))
-);
-
-const makeGetInitialProps = context => ({ component, ...widgetConfig }) =>
-	component.getInitialProps
-		? component.getInitialProps({ ...context, ...widgetConfig })
-		: Promise.resolve(null);
-
-/**
- * Asynchronously returns an array of initial props with the indices corresponding
- * to the passed widget configs.
- *
- * @param {Array} widgetConfigs widget configs
- * @param {Object} context argument to call the initialProps methods with
- */
-const resolveInitialProps = async (context, widgetConfigs) => {
-	const getInitialProps = makeGetInitialProps(context);
-	const promises = R.map(getInitialProps, widgetConfigs);
-	const namespaces = R.map(R.prop('namespace'), widgetConfigs);
-
-	return R.zipObj(namespaces, await Promise.all(promises));
-};
 
 // TODO: similar to getArgValue in `lib/cli.js` of react-union-scripts, we should reuse it somehow
 const getPort = () => {
@@ -62,8 +35,6 @@ const printUsage = port => {
 };
 
 module.exports = {
-	hoistComponentStatics,
-	resolveInitialProps,
 	getPort,
 	printUsage,
 };
